requirements.txt
 opencv-python
streamlit
gtts
genai
os
deep_translator
dotenv
google-geni